Move  profile model, view to user_profile app
