+++ django_project/feti/models/__init__.py
@@ -18,7 +18,6 @@ from feti.models.national_qualifications_framework import (  # noqa
 from feti.models.provider import Provider  # noqa
 from feti.models.learning_pathway import LearningPathway  # noqa
 from feti.models.occupation import Occupation  # noqa
-from feti.models.profile import Profile  # noqa
 
 __author__ = 'Christian Christelis <christian@kartoza.com>'
 __date__ = '04/2015'
