+++ django_project/user_profile/models/__init__.py
@@ -5,3 +5,4 @@ __copyright__ = 'kartoza.com'
 
 from user_profile.models.campus_official import CampusOfficial  # noqa
 from user_profile.models.provider_official import ProviderOfficial  # noqa
+from user_profile.models.profile import Profile  # noqa
