+++ django_project/user_profile/views/__init__.py
@@ -1,0 +1,2 @@
+__author__ = 'Dimas Ciputra <dimas@kartoza.com>'
+__date__ = '22/09/16'
